+++ deployment/k8s/job/create-job.sh
@@ -5,8 +5,8 @@ JAVA_OPTS="-DbaseUrl=http://localhost:8080 -DdurationMin=0.25 -DrequestPerSecond
 SIMULATION_NAME="gatling.test.example.simulation.ExampleGetSimulation"
 
 id=$(uuidgen | cut -c1-8 | awk '{print tolower($0)}')
-JOB_FILENAME=gatling-test-example-job-${id}.yaml
-JOB_NAME=gatling-test-example-${id}
+JOB_FILENAME=gatling-scala-example-job-${id}.yaml
+JOB_NAME=gatling-scala-example-${id}
 ./create-job-yaml.py --out "${JOB_FILENAME}" --name "${JOB_NAME}" --java_opts "${JAVA_OPTS}" --simulation "${SIMULATION_NAME}"
 kubectl apply -f ${JOB_FILENAME}
 echo "JOB_FILENAME=${JOB_FILENAME}" > create-job.vars